Okay, everything's cobra.Command'd, and we share flags now. It still feels like there's some better code structure lurking in there (mostly, we have too much duplication between proxy and fi), but it's not too bad.
There's a better arrangement here - command stuff is in 'cmd', instead of in internal, so now the `internal` stuff feels more a library instead of a CLI-grab-bag.
